Add callbacks to decide if the Sequence will be decoded in hardware or not
Hardware decoding is possible per Sequence, so for each Sequence Header we
ask the host app if it wants to decode it in hardware or in software.

We fallback to software decoding when the host can't decode in hardware.

The data pointers in the Dav1dPicture are custom with hardware decoders and
depend on the type of hardware decoding.

For the moment there is no hardware implementation, but it has been validated
to work with DXVA2, D3D11 and D3D12 with a single decoder thread.

Film-grain is usually not handled by hardware decoders so should not be
applied.

Co-Authored-By: Steve Lhomme <robux4@videolabs.io>