Add support for input from stdin
Allow the user to run dav1d by piping input from stdin.
To do this, specify `-i -` and the stdin will be piped

Since there are multiple areas where a file is opened and the fact that
there is seeking involved, the workaround is to create a temporary file
and put everything from stdin to that temporary file. After that, we
continue processing in the same way as a file and remove the temporary
file at the end.